+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   ([#7944](https://github.com/mitmproxy/mitmproxy/pull/7944), @lups2000)
 - Allow hiding the Quick Help UI in the mitmproxy console with the 'H' key.
   ([#8095](https://github.com/mitmproxy/mitmproxy/pull/8095), @seroperson)
+- Add an additional “Clear All Flows” button in mitmweb for easier access.
+  ([#8121](https://github.com/mitmproxy/mitmproxy/pull/8121), @iBug)
 
 ## 24 November 2025: mitmproxy 12.2.1